OPENMEETINGS-194 --force parameter is added, DB properties are being read from existent config if not specified, NPE in TestSetupCleanupJob is fixed
git-svn-id: https://svn.apache.org/repos/asf/incubator/openmeetings/trunk/singlewebapp@1339461 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/openmeetings/app/quartz/scheduler/TestSetupCleanupJob.java
@@ -35,18 +35,21 @@ public class TestSetupCleanupJob {
 		try {
 			//FIXME need to move all these staff to helper
 			File streams = new File(ScopeApplicationAdapter.batchFileFir);
-			for (File folder : streams.listFiles()) {
-				if (folder.isDirectory()) {
-					//TODO need to rework this and remove hardcodings
-					for (File file : folder.listFiles(new FileFilter() {
-						public boolean accept(File file) {
-							return file.getName().startsWith("TEST_SETUP_");
-						}
-					}))
-					{
-						if (file.isFile() && file.lastModified() + expirationInterval < System.currentTimeMillis()) {
-							log.debug("expired TEST SETUP found: " + file.getAbsolutePath());
-							file.delete();
+			File[] folders = streams.listFiles();
+			if (folders != null) {
+				for (File folder : folders) {
+					if (folder.isDirectory()) {
+						//TODO need to rework this and remove hardcodings
+						for (File file : folder.listFiles(new FileFilter() {
+							public boolean accept(File file) {
+								return file.getName().startsWith("TEST_SETUP_");
+							}
+						}))
+						{
+							if (file.isFile() && file.lastModified() + expirationInterval < System.currentTimeMillis()) {
+								log.debug("expired TEST SETUP found: " + file.getAbsolutePath());
+								file.delete();
+							}
 						}
 					}
 				}
